+++ plat/xilinx/zynqmp/pm_service/pm_ipi.c
@@ -7,6 +7,8 @@
 #include <arch_helpers.h>
 #include <lib/bakery_lock.h>
 #include <lib/mmio.h>
+
+#include <ipi.h>
 #include <plat_ipi.h>
 #include <plat_private.h>
 #include <plat/common/platform.h>
